+++ PPEventRegistryAPI/Classes/API/Operations/PPLoginOperation.swift
@@ -14,15 +14,21 @@ final class PPLoginOperation: PPAsyncOperation {
         super.init(controller: .Login, method: .Post, parameters: parameters)
 
         self.completionHandler = { response, error in
-            if let action = response?["action"] as? String, action == "unknownUser" {
-                DispatchQueue.main.async {
-                    let error = NSError(domain: "Unknown User", code: 0, userInfo: nil)
-                    completionHandler(error)
-                }
-            } else {
+            guard error == nil else {
                 DispatchQueue.main.async {
                     completionHandler(error)
                 }
+                return
+            }
+
+            var credsError: NSError?
+            if let action = response?["action"] as? String,
+                action == "unknownUser" || action == "missingData" {
+                credsError = NSError(domain: action, code: 0, userInfo: nil)
+            }
+
+            DispatchQueue.main.async {
+                completionHandler(credsError)
             }
         }
     }